Remove common voice from evaluation, as discussed.
Pin nemo to a particular version to make sure results are
reproducible. In particular, include:
https://github.com/NVIDIA/NeMo/pull/10054

(Note that the current version of this commit does not do that.)

Make sure that optional dependency cuda-python is included to ensure
that we use cuda graph accelerated decoder inference in RNN-T and TDT
mdoels.