Add UsefulSensors Moonshine benchmark (#43)
* Add UsefulSensors Moonshine benchmark

Due to trainable in-model preprocessor and therefore lack of a
spectrogram preprocessor, we have opted against wrapping the tokenizer
as a processor. Further, we must make substantial changes compared with
existing transformer models, so we decided to create a separate
benchmark.

* Add moonshine-specific requirements.txt.

Adds the `einops` package which our HF hub repo requries.